@@ -1,37 +1,11 @@
 import { StrictMode } from "react";
-import ReactDOM from "react-dom";
-import { createGlobalStyle } from "styled-components";
-import App from "./App";
+import { createRoot } from "react-dom/client";
+import { App } from "./App";
+import { GlobalStyles } from "./styles/GlobalStyles";
 
-const GlobalStyle = createGlobalStyle`
-  * {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    font-family: 'Roboto', sans-serif;
-  }
-  body {
-    background-color: #111;
-    color: #fff;
-  }
-
-  ::-webkit-scrollbar {
-    width: 8px;
-  }
-
-  ::-webkit-scrollbar-track {
-    background-color: #f0f0f0;
-  }
-  
-  ::-webkit-scrollbar-thumb {
-    background-color: #747373;
-  }
-`;
-
-ReactDOM.render(
+createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    <GlobalStyle />
+    <GlobalStyles />
     <App />
-  </StrictMode>,
-  document.getElementById("root")
+  </StrictMode>
 );
