@@ -1,4 +1,4 @@
-import { MovieItem } from "./MovieItem";
+import type { MovieItem } from "./MovieItem";
 
 export type MovieList = {
   slug: string;
